% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_diagnostic.R
\name{berg_envel}
\alias{berg_envel}
\alias{print.berg_envel}
\alias{plot.berg_envel}
\title{Envelope Plot for the BerG Regression Model}
\usage{
berg_envel(object, nsim = 99, progressBar = TRUE, ...)

\method{print}{berg_envel}(x, ...)

\method{plot}{berg_envel}(x, type = c("quantile", "pearson", "response"), level = 0.95, ...)
}
\arguments{
\item{object}{a \code{'bergrm'} object.}

\item{nsim}{the number of replicates. The default is \code{nsim = 99}.}

\item{progressBar}{logical; if \code{TRUE}, a progress bar is displayed giving
the progress of making the graph. It can slow down the function
considerably in applications with a large sample size.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{an \code{'envel_berg'} object.}

\item{type}{character; specifies which residual should be produced in the
envelope plot. The available options are \code{"quantile"} (default),
\code{"pearson"}, and \code{"response"} (raw residuals, y - mu).}

\item{level}{level of the sample quantile of the residual used in the
construction of confidence bands.}
}
\value{
\code{berg_envel} returns an \code{"berg_envel"} object wichi consists of
    a list with the following components:
\describe{
  \item{residuals}{a list with the quantile, pearson, and raw residuals resulting from the
      fit of the BerG regression model.}
  \item{simulation}{a list whose components are matrices containing the
      ordered quantile, pearson, and raw residuals of the simulation for the
      plot envelope.}

 The method \code{plot} makes the envelope plot.
 }
}
\description{
Provides the simulated envelope plot of the residuals (selectable
by \code{type}) resulting from the BerG regression fit.
}
