% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_methods.R
\name{bergreg-methods}
\alias{bergreg-methods}
\alias{print.bergreg}
\alias{summary.bergreg}
\alias{print.summary.bergreg}
\alias{coef.bergreg}
\alias{vcov.bergreg}
\alias{logLik.bergreg}
\alias{AIC.bergreg}
\alias{model.matrix.bergreg}
\title{Methods for 'bergreg' objects}
\usage{
\method{print}{bergreg}(x, ...)

\method{summary}{bergreg}(object, ...)

\method{print}{summary.bergreg}(x, ...)

\method{coef}{bergreg}(object, model = c("full", "mean", "dispersion"), ...)

\method{vcov}{bergreg}(object, model = c("full", "mean", "dispersion"), ...)

\method{logLik}{bergreg}(object, ...)

\method{AIC}{bergreg}(object, ..., k = 2)

\method{model.matrix}{bergreg}(object, model = c("mean", "dispersion"), ...)
}
\arguments{
\item{x, object}{an object of class \code{bergreg}.}

\item{...}{further arguments passed to or from other methods.}

\item{model}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"full"} (default), a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}

\item{k}{numeric, the penalty per parameter to be used; the default
\code{k = 2} is the classical AIC.}
}
\description{
Methods for 'bergreg' objects
}
\author{
Rodrigo M. R. Medeiros <\email{rodrigo.matheus@live.com}>
}
