% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_hypothesis_test.R
\name{disp_test}
\alias{disp_test}
\title{Hypothesis Test for Dispersion Index in the BerG Regression (Under Development)}
\usage{
disp_test(
  object,
  cols = NULL,
  gamma0 = NULL,
  res_start = NULL,
  control = berg_control(...),
  ...
)
}
\arguments{
\item{object}{a 'bergreg' object.}

\item{cols}{an integer vector specifying which regressors must be tested.
If \code{NULL} (default), a constant dispersion test is performed, that is,
all coefficients, with the exception of the intercept, are tested to be
equal to zero.}

\item{gamma0}{vector of the null hypothesis; if \code{NULL} (default) a
vector of zeros is passed to the constant dispersion test.}

\item{res_start}{initial guess for the maximization of the log-likelihood
function restricted to the null hypothesis.}

\item{control}{optimization control parameters.}

\item{...}{additional optimization parameters passed to the function \code{berg_control}.}
}
\value{
It returns a matrix with score, Wald, likelihood ratio, and gradient
 statistics with the respective p-values.
}
\description{
\code{disp_test} performs a hypothesis test on the coefficients associated
    with the dispersion index in the BerG regression.
}
\examples{
\dontrun{
# Loading the 'grazing' dataset
data(grazing)
head(grazing)

# Fitting a double model
# birds ~ BerG(mu_i, phi_i)
# log(mu_i) = beta_0 + beta_1 when_i + beta_2 grazed_i
# log(mu_i) = gamma_0 + gamma_1 when_i + gamma_2 grazed_i

fit <- bergreg(birds ~ when + grazed | when + grazed, grazing)

# Test for constant dispersion
# H0: gamma_1 = gamma_2 = 0
disp_test(fit)
}

}
