% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berg_methods.R
\name{plot.bergreg}
\alias{plot.bergreg}
\title{Diagnostic plots for the BerG Regression}
\usage{
\method{plot}{bergreg}(
  x,
  type = c("quantile", "pearson", "response"),
  which = 1:2,
  ask = prod(graphics::par("mfcol")) < length(which) && grDevices::dev.interactive(),
  ...
)
}
\arguments{
\item{x}{an object of class \code{bergreg}.}

\item{type}{character; specifies which residual should be produced
in the summary plot. The available arguments are "quantile", "pearson",
and "response" (raw residuals, y - mu).}

\item{which}{numeric; if a subset of the plots is required, specify a subset
of the numbers 1:5.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Five plots (selectable by \code{which}) are currently available:
a plot of residuals against fitted values, a plot of residuals against
the indexes, a Normal Q-Q plot, a barplot with comparisons of the
observed and fitted frequencies, and a plot of the sample autocorelations
of the residuals.
}
